initial refactors to generalize
